plant_watering_controller_docker/requirements.txt
# requirement.txt
paho-mqtt==2.1.0
pyyaml==5.4.1
